requirements.txt
# Core
streamlit>=1.30.0
python-dotenv>=1.0.0
requests>=2.31.0

# LangChain ecosystem
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0
langsmith>=0.1.0

# Vector DB
chromadb>=0.4.0
langchain-chroma>=0.1.0

# LLM & Embeddings
openai>=1.0.0
tiktoken>=0.5.0

# Search
tavily-python>=0.3.0

# Document loading
pypdf>=3.0.0
